@@ -80,14 +80,14 @@
     "jotai-tanstack-query": "^0.8.5",
     "lib": "workspace:*",
     "lucide-react": "^0.258.0",
-    "next": "14.2.5",
+    "next": "15.2.4",
     "next-auth": "^4.24.7",
     "next-themes": "^0.2.1",
     "optics-ts": "^2.4.1",
     "protocol": "workspace:*",
-    "react": "18.3.1",
+    "react": "19.1.0",
     "react-day-picker": "^8.10.1",
-    "react-dom": "18.2.0",
+    "react-dom": "19.1.0",
     "react-hook-form": "^7.52.1",
     "react-inlinesvg": "^4.1.3",
     "react-xarrows": "^2.0.2",
@@ -97,8 +97,6 @@
     "tailwindcss-animate": "^1.0.7",
     "ui": "workspace:*",
     "uploadthing": "^7.0.2",
-    "use-immer": "^0.9.0",
-    "wonka": "^6.3.4",
     "zod": "^3.23.8"
   },
   "devDependencies": {
@@ -115,11 +113,11 @@
     "@swc-jotai/react-refresh": "^0.1.1",
     "@tanstack/eslint-plugin-query": "^5.51.15",
     "@types/node": "^22.5.0",
-    "@types/react": "^18.3.4",
-    "@types/react-dom": "^18.3.0",
+    "@types/react": "^19.0.12",
+    "@types/react-dom": "^19.0.4",
     "autoprefixer": "10.4.14",
     "dotenv-cli": "^7.4.2",
-    "eslint": "8.43.0",
+    "eslint": "9.23.0",
     "eslint-config-custom": "workspace:*",
     "eslint-config-next": "14.0.1",
     "eslint-plugin-storybook": "^0.6.15",
