@@ -13,13 +13,14 @@
   "devDependencies": {
     "autoprefixer": "^10.4.19",
     "dotenv-cli": "^7.4.2",
-    "eslint": "^8.57.1",
-    "next": "^14.2.5",
+    "eslint": "^9.23.0",
+    "next": "^15.2.4",
     "postcss": "^8.4.40",
     "prettier": "^3.3.3",
     "tailwindcss": "^3.4.7",
     "tsconfig": "workspace:*",
     "turbo": "^2.1.0"
   },
-  "name": "othi-monorepo"
+  "name": "othi-monorepo",
+  "packageManager": "pnpm@9.13.2"
 }
