@@ -28,7 +28,7 @@
   "devDependencies": {
     "@types/node": "^22.5.0",
     "drizzle-kit": "^0.30.4",
-    "eslint": "^8.57.1",
+    "eslint": "^9.23.0",
     "eslint-config-custom": "workspace:*",
     "eslint-config-next": "14.0.0",
     "typescript": "^5.5.4"
