@@ -3,14 +3,26 @@
   "license": "MIT",
   "version": "0.0.0",
   "private": true,
+    "exports": {
+    "./base": "./base.js",
+    "./next-js": "./next.js",
+    "./react-internal": "./react-internal.js"
+  },
   "devDependencies": {
     "@tanstack/eslint-plugin-query": "^5.51.15",
     "@vercel/style-guide": "^5.2.0",
     "eslint-config-turbo": "^2.1.0",
     "eslint-plugin-storybook": "^0.6.15",
-    "typescript": "^5.5.4"
-  },
-  "dependencies": {
-    "eslint-config-prettier": "^9.1.0"
+    "@eslint/js": "^9.23.0",
+    "@next/eslint-plugin-next": "^15.2.1",
+    "eslint": "^9.23.0",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-plugin-only-warn": "^1.1.0",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-turbo": "^2.4.4",
+    "globals": "^16.0.0",
+    "typescript": "^5.8.2",
+    "typescript-eslint": "^8.27.0"
   }
 }
