@@ -18,10 +18,10 @@
   },
   "devDependencies": {
     "@types/node": "^22.5.0",
-    "@types/react": "^18.3.4",
-    "@types/react-dom": "^18.3.0",
+    "@types/react": "^19.0.12",
+    "@types/react-dom": "^19.0.4",
     "eslint-config-custom": "workspace:*",
-    "react": "^18.3.1",
+    "react": "^19.1.0",
     "tsconfig": "workspace:*",
     "typescript": "^5.5.4"
   }
