@@ -27,9 +27,9 @@
     "env": "workspace:*",
     "jsdom": "^25.0.1",
     "lib": "workspace:*",
-    "next": "^14.2.5",
+    "next": "^15.2.4",
     "next-auth": "^4.24.7",
-    "react": "^18.3.1",
+    "react": "^19.1.0",
     "rehype-stringify": "^10.0.1",
     "remark": "^15.0.1",
     "remark-gfm": "^4.0.1",
@@ -48,7 +48,7 @@
     "@connectrpc/protoc-gen-connect-query": "^1.4.2",
     "@types/dompurify": "^3.0.5",
     "@types/jsdom": "^21.1.7",
-    "@types/react": "^18.3.4",
+    "@types/react": "^19.0.12",
     "@types/turndown": "^5.0.5",
     "eslint-config-custom": "workspace:*",
     "tsconfig": "workspace:*",
