@@ -37,9 +37,9 @@
     "lib": "workspace:*",
     "lucide-react": "^0.292.0",
     "next-themes": "^0.2.1",
-    "react": "^18.3.1",
+    "react": "^19.1.0",
     "react-day-picker": "^8.10.1",
-    "react-dom": "^18.3.1",
+    "react-dom": "^19.1.0",
     "react-hook-form": "^7.52.1",
     "sonner": "^1.5.0",
     "tailwind-merge": "^2.4.0",
@@ -49,8 +49,8 @@
   "devDependencies": {
     "@turbo/gen": "^1.13.4",
     "@types/node": "^22.5.0",
-    "@types/react": "^18.3.4",
-    "@types/react-dom": "^18.3.0",
+    "@types/react": "^19.0.12",
+    "@types/react-dom": "^19.0.4",
     "autoprefixer": "^10.4.19",
     "eslint-config-custom": "workspace:*",
     "postcss": "^8.4.40",
