@@ -17,11 +17,11 @@ const nextConfig = {
   },
   experimental: {
     swcPlugins: [
-      ["@swc-jotai/react-refresh", {}],
-      ["@swc-jotai/debug-label", {}],
+      // ["@swc-jotai/react-refresh", {}],
+      // ["@swc-jotai/debug-label", {}],
     ],
-    serverComponentsExternalPackages: ["@libsql/client"],
   },
+  serverExternalPackages: ["@libsql/client"],
   transpilePackages: ["lib", "ui", "database", "protocol", "auth"],
 };
 
