@@ -1,46 +1,46 @@
-const { resolve } = require("node:path");
+// const { resolve } = require("node:path");
 
-const project = resolve(process.cwd(), "tsconfig.json");
+// const project = resolve(process.cwd(), "tsconfig.json");
 
-/*
- * This is a custom ESLint configuration for use with
- * typescript packages.
- *
- * This config extends the Vercel Engineering Style Guide.
- * For more information, see https://github.com/vercel/style-guide
- *
- */
+// /*
+//  * This is a custom ESLint configuration for use with
+//  * typescript packages.
+//  *
+//  * This config extends the Vercel Engineering Style Guide.
+//  * For more information, see https://github.com/vercel/style-guide
+//  *
+//  */
 
-module.exports = {
-  extends: [
-    "@vercel/style-guide/eslint/node",
-    "@vercel/style-guide/eslint/typescript",
-  ].map(require.resolve),
-  parserOptions: {
-    project,
-    tsconfigRootDir: process.cwd(),
-  },
-  globals: {
-    React: true,
-    JSX: true,
-  },
-  settings: {
-    "import/resolver": {
-      typescript: {
-        project,
-      },
-    },
-  },
-  ignorePatterns: ["node_modules/", "dist/", ".eslintrc.js"],
-  rules: {
-    "import/no-default-export": "off",
-    "import/no-cycle": "off",
-    "unicorn/filename-case": "off",
-    "eslint-comments/require-description": "off",
-    "tsdoc/syntax": "off",
-    "no-console": "warn",
-    "jsx-a11y/heading-has-content": "off",
-    "@typescript-eslint/explicit-function-return-type": "off",
-    "@typescript-eslint/no-empty-interface": "warn",
-  },
-};
+// module.exports = {
+//   extends: [
+//     "@vercel/style-guide/eslint/node",
+//     "@vercel/style-guide/eslint/typescript",
+//   ].map(require.resolve),
+//   parserOptions: {
+//     project,
+//     tsconfigRootDir: process.cwd(),
+//   },
+//   globals: {
+//     React: true,
+//     JSX: true,
+//   },
+//   settings: {
+//     "import/resolver": {
+//       typescript: {
+//         project,
+//       },
+//     },
+//   },
+//   ignorePatterns: ["node_modules/", "dist/", ".eslintrc.js"],
+//   rules: {
+//     "import/no-default-export": "off",
+//     "import/no-cycle": "off",
+//     "unicorn/filename-case": "off",
+//     "eslint-comments/require-description": "off",
+//     "tsdoc/syntax": "off",
+//     "no-console": "warn",
+//     "jsx-a11y/heading-has-content": "off",
+//     "@typescript-eslint/explicit-function-return-type": "off",
+//     "@typescript-eslint/no-empty-interface": "warn",
+//   },
+// };
